requirements.txt
# Step 1: Update system packages
sudo apt update && sudo apt upgrade -y

# Step 2: Install system dependencies (very important)
sudo apt install -y \
    python3-pip python3-venv \
    build-essential cmake git \
    libatlas-base-dev libopenblas-dev \
    libjpeg-dev libpng-dev libtiff-dev \
    libavcodec-dev libavformat-dev libswscale-dev \
    libgtk-3-dev \
    python3-opencv

# Step 3: Create project folder
mkdir -p ~/Arclight
cd ~/Arclight

# Step 4: Create virtual environment
python3 -m venv arclight_env

# Step 5: Activate virtual environment
source arclight_env/bin/activate

# Step 6: Create requirements.txt (copy this block)
cat << EOF > requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
pandas
insightface
onnxruntime
pillow
EOF

# Step 7: Install Python packages
pip install -r requirements.txt

# Step 8: Check if everything installed successfully
python -c "import cv2, numpy, fastapi, insightface; print('✅ All core packages installed successfully!')"